isMarkerInsideKiruna is now a util instead of a context
